+++ test/languages_test.js
@@ -8,6 +8,9 @@ tape.test('verify language files structure', function(assert) {
     var translations = languages.instructions;
     var english = translations.en;
 
+    assert.ok(english.v5.depart.default.namedistance, 'en.js has depart namedistance');
+    assert.ok(english.v5.continue.straight.namedistance, 'en.js has continue straight namedistance');
+
     Object.keys(translations).forEach((l) => {
         if (l === 'en') return; // do not need to compare to self
         var translation = translations[l];
@@ -23,7 +26,7 @@ tape.test('verify language files structure', function(assert) {
             assert.deepEqual(
                 Object.keys(translation.v5.constants[c]),
                 Object.keys(english.v5.constants[c]),
-                l + ' has correct contant ' + c + ' keys');
+                l + ' has correct constant ' + c + ' keys');
         });
 
         assert.deepEqual(
@@ -31,6 +34,22 @@ tape.test('verify language files structure', function(assert) {
             Object.keys(english.v5.rotary.default),
             l + ' has correct rotary variance keys'
         );
+
+        assert.deepEqual(
+            Object.keys(translation.v5.depart.default),
+            Object.keys(english.v5.depart.default),
+            l + ' has correct depart namedistance keys'
+        );
+
+        // exclude zh-Hans until full translation is available
+        if (l !== 'zh-Hans') {
+            assert.deepEqual(
+                Object.keys(translation.v5.continue.straight),
+                Object.keys(english.v5.continue.straight),
+                l + ' has correct continue straight namedistance keys'
+            );
+        }
+
     });
 
     assert.end();