+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. For change 
 
 - Added a `formatToken` option to `compile`, `getWayName`, and `tokenize` that allows you to manipulate any token value after any grammar or capitalization rules are applied but before the value is inserted into the instruction. [#170](https://github.com/Project-OSRM/osrm-text-instructions/pull/170)
 - Removed the `options` parameter to this module, including the `tokenizedInstruction` hook. Use the `formatToken` option instead. [#170](https://github.com/Project-OSRM/osrm-text-instructions/pull/170)
+- Added `namedistance` option to the depart instructions and added tests for `namedistance` in depart and continue instructions. Some typo fixes. [#177](https://github.com/Project-OSRM/osrm-text-instructions/pull/177#discussion_r145279715)
 
 ## 0.9.0 2017-10-05